Create COVID 19 portfolio
This is a COVID 19 report dating from January 2020 to April 2021. It shows the total deaths, total cases, and other insights. The purpose of this report was to provide insights such as death percentage, death rates, and infection rates, these helps us to better understand the impact of the COVID 19 virus during this period. The is mainly a description analysis with the aim of broadening our understanding of the impact of the COVID 19 pandemic.